refactor: 리더보드 텍스트가 덜 잘리도록 수정
