--- src/app/(auth-required)/leaderboards/components/index.ts
@@ -1,2 +1,0 @@
-export * from './Ranker';
-export * from './Rank';
